requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
flake8==6.0.0
idna==3.3
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pyflakes==3.0.1
pyTelegramBotAPI==4.6.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
schedule==1.1.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.11
